requirements.txt
google-generativeai
anthropic
python-dotenv
ipython
